pyproject.toml
[project]
name = "tg-bot-analyst-moderation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiogram>=3.20.0.post0",
    "alembic>=1.15.2",
    "apscheduler>=3.11.0",
    "asyncpg>=0.30.0",
    "cachetools>=5.5.2",
    "cryptography>=46.0.3",
    "fastapi>=0.118.3",
    "openrouter>=0.1.1",
    "psycopg2-binary>=2.9.10",
    "punq>=0.7.0",
    "pydantic-settings>=2.9.1",
    "pyjwt[crypto]>=2.10.1",
    "pytz>=2025.2",
    "redis[asyncio]>=6.4.0",
    "sqlalchemy[asyncio]>=2.0.40",
    "taskiq-redis>=1.2.0",
    "uvicorn>=0.37.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "ruff>=0.14.5",
    "vulture>=2.14",
]

[tool.mypy]
python_version = "3.10"
files = ["src"]
mypy_path = "src"
plugins = ["pydantic.mypy"]
strict = true
ignore_missing_imports = true
explicit_package_bases = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
